update overview file’s markup
i don’t think this will solve all the invalid <p> wrappings that are happening when the markdown is converted to HTML, but this should hopefully fix a few of them.

Removed <p>’s that were wrapping blockquotes and unordered lists, as those are not valid children of paragraph elements.