no-js fix
there was nothing in place to remove the no-js class that was on the HTML element.

the added script will remove the no-js class and add a js class in its place.

this script is purposefully placed in the head as a stand alone script, before the CSS, so as to ensure that there are no “styling flashes” of no-js to js styled content.  this script needs to be by itself and located before the css for the style flash fix to work, which is why it’s not included in the JS area at the bottom of the file.