issue 86: Jump to top button
initial work for jump to top button functionality.

adds <a> markup to index.twig file that points to the <main> container ID.

adds jump-link-fix script to ensure that browsers that don’t properly set focus to unfocusable elements, will add a tabindex -1 to do so.

back-to-top.js adds in scroll handler script and functionality to show/hide the back to top link only when the page is scrolled a certain amount.

kss.css is the temporary home for the (horribly styled) back to top link.  please make this so much better :)