site logo css updates
noticed that the logo w/text example had the text hidden.  so removed the styling on the h1 that duplicated the visuallyhidden class.

while cleaning that up, i realized that the site logo class must have originally been on an <a> element, as the styles weren’t really doing anything for the <div>s they were now associated with.  so cleaned up that styling and re-implemented some basic hover/focus states for the <a>s inside the site-logo container.