updates
scss:
add some more info to the navbar mobile icon lead-in bullet points, concerning placement within a navigation for landmark discovery.

also slight wording change for the bullet about alt text, as it read to me that no alt=“” is needed, when i know that’s not what it was trying to say.

twig:
add aria-controls attributes to the buttons.  While only JAWS currently does a good job at supporting aria-controls, is an attribute that __should__ have better support, and it can only help in adding additional context to what the button should be doing.